@zhaoyuchen2018 2020-01-08
> 除了`ParamConfig` -> `Config`外，我没有别的大的修改意见。我认为`Config`一个词足以表达意思，没有必要堆叠两个。这个可以请别的同学看下。
> 
> 另外，gpu launch config这个是放在platform下面，目的是给CUDA kernel启动提供一个通用的配置接口，那么下面两个方面的配置还是要考虑的，只是可以后面再加：
> 
> * 允许用户配置线程数
> * 线程数对齐到wrarp或者half-warp。

命名上的我觉得大家都能理解吧，没有必要一定修改。
其他的参数配置方案后面根据需求再添加。